Added FieldInitializer
