+++ CMakeLists.txt
@@ -26,7 +26,7 @@ if (BUILD_TTTCORE STREQUAL "PREBUILT")
       )
   endif()
 else()
-  set(core_src src/core/event.cpp src/core/game.cpp src/core/state.cpp)
+  set(core_src src/core/event.cpp src/core/game.cpp src/core/state.cpp src/core/field.cpp)
   if (BUILD_TTTCORE STREQUAL "FULL")
     set(core_src ${core_src} ../baseline.cpp)
   endif()
