+++ src/core/game.hpp
@@ -45,7 +45,7 @@ class Game {
   State m_state;
 
 public:
-  Game(const State::Opts &opts);
+  Game(const State::Opts &opts, const IFieldInitializer* initializer = nullptr);
   Game(const State &state);
 
   const State &get_state() const;
@@ -55,6 +55,8 @@ public:
   IPlayer *remove_player(Sign sign);
   void add_observer(IObserver *observer);
   void remove_observer(IObserver *observer);
+
+  void set_field_initializer(const IFieldInitializer* initializer);
 
   MoveResult process();
   void reset();
