+++ src/core/state.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include "field.hpp"
 
 namespace ttt::game {
 
@@ -27,56 +28,6 @@ enum class Sign {
   WALL
 };
 
-class Obstacle {
-  char* m_moves;
-  int m_left_size;
-  int m_right_size;
-  int m_up_size;
-  int m_down_size;
-public:
-  Obstacle(int seq_len);
-  Obstacle(const Obstacle& other) = delete;
-  Obstacle& operator=(const Obstacle& other) = delete;
-  int get_rsize() const {return m_right_size;}
-  int get_lsize() const {return m_left_size;}
-  int get_usize() const {return m_up_size;}
-  int get_dsize() const {return m_down_size;}
-  char get_move(int i) const {return m_moves[i];};
-  int get_moves_len() const;
-  ~Obstacle();
-};
-
-class FieldBitmap {
-  char *m_bitmap;
-  int m_rows;
-  int m_cols;
-  float m_playable_part;
-  int m_wall_n;
-  int m_max_obstacle_len;
-public:
-  FieldBitmap(int rows, int cols, float playable_part, int max_obstacle_len);
-  FieldBitmap(const FieldBitmap &other);
-  FieldBitmap(FieldBitmap &&other);
-  ~FieldBitmap();
-
-  FieldBitmap &operator=(const FieldBitmap &other);
-  FieldBitmap &operator=(FieldBitmap &&other);
-
-  void set(int x, int y, Sign s);
-  
-  void generate();
-
-  Sign get(int x, int y) const;
-  bool is_valid(int x, int y) const;
-  int get_wall_cells_num() const {return m_wall_n;};
-
-private:
-  int _bitmap_size() const;
-  void _set_unsafe(int x, int y, Sign s);
-  void _find_obstacle_place(const Obstacle& obstacle, int& x, int& y) const;
-  int _insert_obstacle(const Obstacle& obstacle, int x, int y, int max_len);
-};
-
 class State {
 public:
   struct Opts {
@@ -90,7 +41,7 @@ public:
 
 private:
   Opts m_opts;
-
+  IFieldInitializer* m_initializer;
   FieldBitmap m_field;
   int m_move_no;
   Status m_status;
@@ -98,9 +49,9 @@ private:
   Sign m_winner;
 
 public:
-  State(const Opts &opts);
-  State(const State &state) = default;
-  ~State() = default;
+  State(const Opts &opts, const IFieldInitializer* initializer = nullptr);
+  State(const State &state);
+  ~State();
 
   void reset();
   MoveResult process_move(Sign player, int x, int y);
@@ -114,12 +65,13 @@ public:
 
   State &operator=(const State &state) = default;
 
+  void set_field_initializer(const IFieldInitializer* initializer);
 
 private:
   bool _valid_coords(int x, int y) const;
   void _set_value(int x, int y, Sign sign);
   Sign _opp_sign(Sign player);
   bool _is_winning(int x, int y);
-  void _empty_state();
+  void _reset_state();
 };
 }; // namespace ttt::game
