+++ tests/test_baseline.cpp
@@ -2,6 +2,7 @@
 #include "player/my_observer.hpp"
 #include "player/my_player.hpp"
 #include "core/game.hpp"
+#include "core/field.hpp"
 
 
 #include <cstdio>
@@ -18,8 +19,8 @@ int main(int argc, char *argv[]) {
   opts.rows = opts.cols = 15;
   opts.win_len = 5;
   opts.max_moves = 0;
-  opts.playable_part = 0.66;
-  opts.max_obstacle_len = 3;
+  
+  auto field_initializer = ttt::game::ObstaclesFieldInitializer(0.66, 3);
 
   ttt::game::IPlayer *p1 = new ttt::my_player::MyPlayer("MyPlayer");  // <~~ add your player here
   ttt::game::IPlayer *p2 = ttt::baseline::get_harder_player("p_easy");
@@ -28,7 +29,7 @@ int main(int argc, char *argv[]) {
 
   ttt::my_player::ConsoleWriter obs;
 
-  ttt::game::Game game(opts);
+  ttt::game::Game game(opts, &field_initializer);
   game.add_player(ttt::game::Sign::X, p1);
   game.add_player(ttt::game::Sign::O, p2);
   game.add_observer(&obs);
